requirements.txt
django
flask
djangorestframework
django-cors-headers
requests
